+++ buttons.php
@@ -40,50 +40,15 @@ $csrf = htmlspecialchars($_SESSION['csrf_token'], ENT_QUOTES, 'UTF-8');
 
 $allowed_styles = ['btn-default', 'btn-primary', 'btn-success', 'btn-info', 'btn-warning', 'btn-danger'];
 $allowed_sizes  = ['btn-xs', 'btn-sm', 'btn-md', 'btn-lg'];
+
+$page_title = 'Buttons';
+require_once('./include/header.php');
 ?>
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="description" content="GumCP Command Buttons">
-    <link rel="shortcut icon" href="./static/images/raspberry.png" type="image/png" />
-    <link rel="icon" href="./static/images/raspberry.png" type="image/png" />
-    <title>GumCP Buttons</title>
-    <link href="./static/css.php" rel="stylesheet" type="text/css">
-    <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
-    <script src="./static/js.php" type="text/javascript"></script>
-    <script>
+
+<script>
     var CSRF_TOKEN         = <?php echo json_encode($_SESSION['csrf_token']); ?>;
     var BUTTON_API_ENABLED = <?php echo json_encode(!empty($gumcp_modules['button_api']['module_active'])); ?>;
     </script>
-</head>
-
-<body>
-<div class="container">
-
-    <nav class="navbar navbar-default">
-        <div class="container-fluid">
-            <div class="navbar-header">
-                <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-                    <span class="sr-only">Toggle navigation</span>
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                </button>
-                <a class="navbar-brand" href="./index.php">
-                    <img src="./static/images/raspberry.png" alt="Logo" />GumCP
-                </a>
-            </div>
-            <div id="navbar" class="navbar-collapse collapse">
-                <ul class="nav navbar-nav navbar-right">
-                    <?php require_once('./include/menu.php'); ?>
-                </ul>
-            </div>
-        </div>
-    </nav>
-
     <div class="page-header">
         <h1><?php echo htmlspecialchars(t('btn.title', 'Command Buttons'), ENT_QUOTES, 'UTF-8'); ?></h1>
     </div>
@@ -366,16 +331,4 @@ $allowed_sizes  = ['btn-xs', 'btn-sm', 'btn-md', 'btn-lg'];
     </div>
 </div>
 
-<footer class="footer">
-    <div class="container">
-        <p class="text-muted">
-            GumCP <a href="https://github.com/gumslone/GumCP" target="_blank" rel="noopener">GitHub</a>.
-            <a href="https://www.paypal.com/donate/?hosted_button_id=VCWHQPACTXV5N" target="_blank" rel="noopener">
-                <img src="./static/images/Donate-PayPal-green.svg" alt="Donate"/>
-            </a>
-        </p>
-    </div>
-</footer>
-
-</body>
-</html>
+<?php require_once('./include/footer.php'); ?>