+++ cron.php
@@ -8,51 +8,17 @@ require_once('./include/init.php');
 if (!isset($_SESSION['csrf_token'])) {
     $_SESSION['csrf_token'] = bin2hex(random_bytes(32));
 }
+
+$page_title = 'Cron Jobs';
+require_once('./include/header.php');
 ?>
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="description" content="GumCP Cron Jobs">
-    <link rel="shortcut icon" href="./static/images/raspberry.png" type="image/png">
-    <link rel="icon"          href="./static/images/raspberry.png" type="image/png">
-    <title>GumCP Cron Jobs</title>
-    <link href="./static/css.php" rel="stylesheet" type="text/css">
-    <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
-    <script src="./static/js.php" type="text/javascript"></script>
-    <script>
+
+<script>
     var CSRF_TOKEN = <?php echo json_encode($_SESSION['csrf_token']); ?>;
     var T_CRON_NONE    = <?php echo json_encode(t('cron.none', 'No cron jobs for this user.')); ?>;
     var T_CRON_CUSTOM  = <?php echo json_encode(t('cron.custom', 'Custom schedule')); ?>;
     var T_CRON_INVALID = <?php echo json_encode(t('cron.invalid', 'Invalid schedule expression.')); ?>;
     </script>
-</head>
-
-<body>
-<div class="container">
-
-    <nav class="navbar navbar-default">
-        <div class="container-fluid">
-            <div class="navbar-header">
-                <button type="button" class="navbar-toggle collapsed" data-toggle="collapse"
-                        data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-                    <span class="sr-only">Toggle navigation</span>
-                    <span class="icon-bar"></span><span class="icon-bar"></span><span class="icon-bar"></span>
-                </button>
-                <a class="navbar-brand" href="./index.php">
-                    <img src="./static/images/raspberry.png" alt="Logo"> GumCP
-                </a>
-            </div>
-            <div id="navbar" class="navbar-collapse collapse">
-                <ul class="nav navbar-nav navbar-right">
-                    <?php require_once('./include/menu.php'); ?>
-                </ul>
-            </div>
-        </div>
-    </nav>
-
     <div class="page-header">
         <h1><i class="fa fa-clock-o"></i> <?php echo htmlspecialchars(t('cron.title', 'Cron Jobs'), ENT_QUOTES, 'UTF-8'); ?></h1>
     </div>
@@ -119,18 +85,6 @@ if (!isset($_SESSION['csrf_token'])) {
     </div>
 
 </div>
-
-<footer class="footer">
-    <div class="container">
-        <p class="text-muted">
-            GumCP <a href="https://github.com/gumslone/GumCP" target="_blank" rel="noopener">GitHub</a>.
-            <a href="https://www.paypal.com/donate/?hosted_button_id=VCWHQPACTXV5N"
-               target="_blank" rel="noopener">
-                <img src="./static/images/Donate-PayPal-green.svg" alt="Donate">
-            </a>
-        </p>
-    </div>
-</footer>
 
 <script>
 function esc(s) { return $('<div>').text(s == null ? '' : s).html(); }
@@ -335,5 +289,4 @@ function cronDelete(idx) {
 $(cronLoad);
 </script>
 
-</body>
-</html>
+<?php require_once('./include/footer.php'); ?>