+++ phpinfo.php
@@ -14,21 +14,12 @@ $phpinfo = '';
 if (preg_match('/<body[^>]*>(.*)<\/body>/isU', $html, $matches)) {
     $phpinfo = $matches[1];
 }
+
+$page_title = 'PHP Info';
+require_once('./include/header.php');
 ?>
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="description" content="GumCP PHP Info">
-    <link rel="shortcut icon" href="./static/images/raspberry.png" type="image/png">
-    <link rel="icon"          href="./static/images/raspberry.png" type="image/png">
-    <title>GumCP PHP Info</title>
-    <link href="./static/css.php" rel="stylesheet" type="text/css">
-    <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
-    <script src="./static/js.php" type="text/javascript"></script>
-    <style>
+
+<style>
         /* Scope phpinfo() styles to avoid polluting Bootstrap */
         #phpinfo-output pre { margin:0; font-family:monospace; }
         #phpinfo-output a:link  { color:#009; text-decoration:none; background-color:#fff; }
@@ -48,33 +39,6 @@ if (preg_match('/<body[^>]*>(.*)<\/body>/isU', $html, $matches)) {
         #phpinfo-output img { float:right; border:0; }
         #phpinfo-output hr  { width:934px; background-color:#ccc; border:0; height:1px; }
     </style>
-</head>
-
-<body>
-<div class="container">
-
-    <nav class="navbar navbar-default">
-        <div class="container-fluid">
-            <div class="navbar-header">
-                <button type="button" class="navbar-toggle collapsed" data-toggle="collapse"
-                        data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-                    <span class="sr-only">Toggle navigation</span>
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                </button>
-                <a class="navbar-brand" href="./index.php">
-                    <img src="./static/images/raspberry.png" alt="Logo"> GumCP
-                </a>
-            </div>
-            <div id="navbar" class="navbar-collapse collapse">
-                <ul class="nav navbar-nav navbar-right">
-                    <?php require_once('./include/menu.php'); ?>
-                </ul>
-            </div>
-        </div>
-    </nav>
-
     <div class="panel panel-default" style="margin-bottom:5px">
         <div class="panel-heading">
             <h3 class="panel-title"><i class="fa fa-info-circle"></i> PHP Info</h3>
@@ -93,17 +57,4 @@ if (preg_match('/<body[^>]*>(.*)<\/body>/isU', $html, $matches)) {
 
 </div>
 
-<footer class="footer">
-    <div class="container">
-        <p class="text-muted">
-            GumCP <a href="https://github.com/gumslone/GumCP" target="_blank" rel="noopener">GitHub</a>.
-            <a href="https://www.paypal.com/donate/?hosted_button_id=VCWHQPACTXV5N"
-               target="_blank" rel="noopener">
-                <img src="./static/images/Donate-PayPal-green.svg" alt="Donate">
-            </a>
-        </p>
-    </div>
-</footer>
-
-</body>
-</html>
+<?php require_once('./include/footer.php'); ?>