+++ rpi.php
@@ -17,46 +17,12 @@ $interfaces = [
     'vnc'     => 'VNC server',
     'camera'  => 'Camera (legacy)',
 ];
+
+$page_title = 'Raspberry Pi';
+require_once('./include/header.php');
 ?>
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="description" content="GumCP Raspberry Pi tools">
-    <link rel="shortcut icon" href="./static/images/raspberry.png" type="image/png">
-    <link rel="icon"          href="./static/images/raspberry.png" type="image/png">
-    <title>GumCP Raspberry Pi</title>
-    <link href="./static/css.php" rel="stylesheet" type="text/css">
-    <link href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
-    <script src="./static/js.php" type="text/javascript"></script>
-    <script>var CSRF_TOKEN = <?php echo json_encode($_SESSION['csrf_token']); ?>;</script>
-</head>
 
-<body>
-<div class="container">
-
-    <nav class="navbar navbar-default">
-        <div class="container-fluid">
-            <div class="navbar-header">
-                <button type="button" class="navbar-toggle collapsed" data-toggle="collapse"
-                        data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-                    <span class="sr-only">Toggle navigation</span>
-                    <span class="icon-bar"></span><span class="icon-bar"></span><span class="icon-bar"></span>
-                </button>
-                <a class="navbar-brand" href="./index.php">
-                    <img src="./static/images/raspberry.png" alt="Logo"> GumCP
-                </a>
-            </div>
-            <div id="navbar" class="navbar-collapse collapse">
-                <ul class="nav navbar-nav navbar-right">
-                    <?php require_once('./include/menu.php'); ?>
-                </ul>
-            </div>
-        </div>
-    </nav>
-
+<script>var CSRF_TOKEN = <?php echo json_encode($_SESSION['csrf_token']); ?>;</script>
     <div class="page-header">
         <h1><i class="fa fa-microchip"></i> <?php echo htmlspecialchars(t('nav.rpi', 'Raspberry Pi'), ENT_QUOTES, 'UTF-8'); ?></h1>
     </div>
@@ -151,18 +117,6 @@ $interfaces = [
     </div>
 
 </div>
-
-<footer class="footer">
-    <div class="container">
-        <p class="text-muted">
-            GumCP <a href="https://github.com/gumslone/GumCP" target="_blank" rel="noopener">GitHub</a>.
-            <a href="https://www.paypal.com/donate/?hosted_button_id=VCWHQPACTXV5N"
-               target="_blank" rel="noopener">
-                <img src="./static/images/Donate-PayPal-green.svg" alt="Donate">
-            </a>
-        </p>
-    </div>
-</footer>
 
 <script>
 function esc(s) { return $('<div>').text(s == null ? '' : s).html(); }
@@ -343,5 +297,4 @@ $(function() {
 });
 </script>
 
-</body>
-</html>
+<?php require_once('./include/footer.php'); ?>